Version 3
This version was created to create the directories needed to recompile wim into the ISO. When shared at the office, they would copy the script and not the directories needed to have the script work. So I built in to the script to create the directories. I am still working on this version to improve it